steam floats usage reduced - change to big decimals

+++ app/model/Rates.scala
@@ -2,8 +2,8 @@ package model
 
 import scala.xml.XML
 
-case class Rates(euroTo: Map[String, Float]) {
-  def recalculateFromEuro(valueInEuro: Float, currency: String): Option[Float] = {
+case class Rates(euroTo: Map[String, BigDecimal]) {
+  def recalculateFromEuro(valueInEuro: BigDecimal, currency: String): Option[BigDecimal] = {
     euroTo.get(currency).map(_ * valueInEuro)
   }
 }
@@ -12,6 +12,6 @@ object Rates {
   def parseFromXml(xml: String): Rates = {
     val parsed = XML.loadString(xml)
     val rateMaps = (parsed \ "Cube" \ "Cube" \ "Cube").map(_.attributes.asAttrMap)
-    Rates(rateMaps.map(e => (e("currency"), e("rate").toFloat)).toMap)
+    Rates(rateMaps.map(e => (e("currency"), BigDecimal(e("rate")))).toMap)
   }
 }
